[fix] Fix the build errors in next.js
