+++ src/app/[[...slug]]/client.tsx
@@ -1,7 +1,9 @@
 'use client';
 import dynamic from 'next/dynamic';
 
-const App = dynamic(() => import('../../App').then((module) => module.App));
+const App = dynamic(() => import('../../App').then((module) => module.App), {
+  ssr: false,
+});
 
 export function ClientOnly() {
   return <App />;
