+++ src/app_pages/Error/index.tsx
@@ -1,0 +1,11 @@
+import { Layout } from '@/components/styles/Layout';
+
+export const NotFoundPage = () => {
+  return (
+    <Layout>
+      <div>존재하지 않는 페이지입니다.</div>
+    </Layout>
+  );
+};
+
+export default NotFoundPage;
