+++ src/app_pages/SignUp/index.tsx
@@ -1,0 +1,11 @@
+import { Layout } from '@/components/styles/Layout';
+
+export const SignUpPage = () => {
+  return (
+    <Layout>
+      <div>회원가입페이지</div>
+    </Layout>
+  );
+};
+
+export default SignUpPage;
