--- src/pages/Error/index.tsx
@@ -1,9 +1,0 @@
-import { Layout } from '@/components/styles/Layout';
-
-export const NotFoundPage = () => {
-  return (
-    <Layout>
-      <div>존재하지 않는 페이지입니다.</div>
-    </Layout>
-  );
-};
