--- src/pages/SignUp/index.tsx
@@ -1,9 +1,0 @@
-import { Layout } from '@/components/styles/Layout';
-
-export const SignUpPage = () => {
-  return (
-    <Layout>
-      <div>회원가입페이지</div>
-    </Layout>
-  );
-};
